Move guiTests to another folder
