+++ src/test/java/seedu/address/ui/guitests/AppParametersTest.java
@@ -1,4 +1,4 @@
-package seedu.address;
+package seedu.address.ui.guitests;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -13,6 +13,7 @@ import java.util.Map;
 import org.junit.jupiter.api.Test;
 
 import javafx.application.Application;
+import seedu.address.AppParameters;
 
 public class AppParametersTest {
