+++ src/test/java/seedu/address/ui/guitests/GuiUnitTest.java
@@ -1,4 +1,4 @@
-package seedu.address.ui;
+package seedu.address.ui.guitests;
 
 import java.util.Optional;
