+++ src/test/java/seedu/address/ui/guitests/TestFxmlObject.java
@@ -1,4 +1,4 @@
-package seedu.address.ui;
+package seedu.address.ui.guitests;
 
 import java.util.Objects;
