+++ src/test/resources/view/UiPartTest/validFileWithFxRoot.fxml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<fx:root type="seedu.address.ui.TestFxmlObject" xmlns="http://javafx.com/javafx/17"
-        xmlns:fx="http://javafx.com/fxml/1">
+<fx:root type="TestFxmlObject" xmlns="http://javafx.com/javafx/17"
+         xmlns:fx="http://javafx.com/fxml/1">
     <text>Hello World!</text>
 </fx:root>
